Corrected websocket_client connections and deprecated renderer.js
