+++ electron-frontend/package.json
@@ -16,8 +16,8 @@
     "styled-components": "^5.3.10"
   },
   "devDependencies": {
+    "@vitejs/plugin-react": "^4.0.0",
     "typescript": "^5.0.0",
-    "vite": "^5.0.0",
-    "@vitejs/plugin-react": "^4.0.0"
+    "vite": "^6.3.5"
   }
 }
